refactor: extract HeritageMap components for better separation of concerns (Phase 4 - part 1)
**Created:**
- src/components/Map/MapHelperComponents.tsx (101 lines)
  - MapCenterHandler - Smooth pan to highlighted sites
  - ZoomLogger - Debug logging for zoom level
  - ScrollWheelHandler - Ctrl+Scroll zoom behavior

- src/components/Map/MapTileLayers.tsx (35 lines)
  - Configures LayersControl with Street/Satellite toggle
  - Encapsulates tile configuration logic
  - Uses useTileConfig hook for Arabic/English tiles

- src/components/Map/MapMarkers.tsx (83 lines)
  - Renders all heritage site markers
  - Handles highlighted (teardrop) vs default (circle) markers
  - Color-coded by status with popup support

**Modified:**
- src/components/Map/HeritageMap.tsx (245 → 84 lines, 66% reduction)
  - Clean composition using extracted components
  - Focused on glow calculations and map container
  - Much easier to read and maintain

**Benefits:**
- Map helper components now independently testable
- Tile and marker logic separated for clarity
- Easy to add new map features without cluttering main component
- Better code organization for future timeline animation work

**Testing:**
- All 184 tests passing
- Linter clean (0 errors, 0 warnings)
- App verified working in browser
